fixes #5

+++ src/main/groovy/com/github/lburgazzoli/gradle/plugin/karaf/kar/KarafKarTask.groovy
@@ -85,7 +85,9 @@ class KarafKarTask extends Jar {
                 Files.createDirectories(destination.parent)
             }
 
-            Files.copy(source, destination)
+            if (!Files.exists(destination)) {
+                Files.copy(source, destination)
+            }
         }
     }
 }
